+++ Source/DungeonGenerator/Private/Parameter/DungeonRoomMeshSetDatabase.cpp
@@ -11,13 +11,13 @@ All Rights Reserved.
 #include "Helper/DungeonDebugUtility.h"
 #endif
 
-const FDungeonMeshSet* UDungeonRoomMeshSetDatabase::AtImplement(const size_t index) const
+const FDungeonTemporaryMeshSet* UDungeonRoomMeshSetDatabase::AtImplement(const size_t index) const
 {
 	const int32 size = Parts.Num();
 	return (size > 0) ? &Parts[index % size] : nullptr;
 }
 
-const FDungeonMeshSet* UDungeonRoomMeshSetDatabase::SelectImplement(const std::shared_ptr<dungeon::Random>& random) const
+const FDungeonTemporaryMeshSet* UDungeonRoomMeshSetDatabase::SelectImplement(const std::shared_ptr<dungeon::Random>& random) const
 {
 	const int32 size = Parts.Num();
 	if (size <= 0)