+++ Source/DungeonGenerator/Private/Parameter/DungeonTemporaryMeshSetDatabase.cpp
@@ -1,0 +1,7 @@
+/**
+@author		Shun Moriya
+@copyright	2023- Shun Moriya
+All Rights Reserved.
+*/
+
+#include "Parameter/DungeonTemporaryMeshSetDatabase.h"