+++ Source/DungeonGeneratorEditor/Private/Parameter/DungeonMeshSetDatabaseTypeActions.cpp
@@ -1,0 +1,28 @@
+/**
+@author		Shun Moriya
+@copyright	2024- Shun Moriya
+All Rights Reserved.
+*/
+
+#include "Parameter/DungeonMeshSetDatabaseTypeActions.h"
+#include "Parameter/DungeonMeshSetDatabase.h"
+
+FDungeonMeshSetDatabaseTypeActions::FDungeonMeshSetDatabaseTypeActions(EAssetTypeCategories::Type InAssetCategory)
+	: mAssetCategory(InAssetCategory)
+{
+}
+
+FText FDungeonMeshSetDatabaseTypeActions::GetName() const
+{
+	return FText::FromName(TEXT("Mesh set database"));
+}
+
+UClass* FDungeonMeshSetDatabaseTypeActions::GetSupportedClass() const
+{
+	return UDungeonMeshSetDatabase::StaticClass();
+}
+
+uint32 FDungeonMeshSetDatabaseTypeActions::GetCategories()
+{
+	return mAssetCategory;
+}